Working gui version.
Added hideable window interface.
Second precision in static_time_window is now decided in runtime.

+++ inc/window.hpp
@@ -25,9 +25,9 @@ enum class Message {
 };
 
 struct coord {
-	uint32_t x;
-	uint32_t y;
-	coord(uint32_t x, uint32_t y);
+	uint16_t x;
+	uint16_t y;
+	coord(uint16_t x, uint16_t y);
 };
 
 bool operator<(const coord &c1, const coord &c2);
